@@ -36,16 +36,16 @@
   ],
   "devDependencies": {
     "@seneca/gateway": "^0.1.5",
-    "@seneca/maintain": "^0.0.19",
-    "@types/jest": "^27.0.1",
-    "esbuild": "^0.12.25",
+    "@seneca/maintain": "^0.0.21",
+    "@types/jest": "^27.5.0",
+    "esbuild": "^0.14.38",
     "esbuild-jest": "^0.5.0",
     "express": "^4.17.1",
-    "jest": "^27.1.0",
+    "jest": "^28.1.0",
     "seneca": "^3.23.3",
     "seneca-msg-test": "^3.2.0",
     "seneca-promisify": "^2.0.0",
-    "typescript": "^4.4.2"
+    "typescript": "^4.6.4"
   },
   "peerDependencies": {
     "@seneca/gateway": ">=0",
